requirements.txt
numpy==2.2.0
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
tzdata==2024.2
Unidecode==1.3.8
